+++ bundles/CoreBundle/src/Migrations/Version20220906111031.php
@@ -29,11 +29,15 @@ final class Version20220906111031 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE `recyclebin` ADD INDEX `recyclebin_date` (`date`)');
+        if($schema->hasTable('recyclebin') && !$schema->getTable('recyclebin')->hasIndex('recyclebin_date')) {
+            $this->addSql('ALTER TABLE `recyclebin` ADD INDEX `recyclebin_date` (`date`)');
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE `recyclebin` DROP INDEX `recyclebin_date`');
+        if($schema->hasTable('recyclebin') && $schema->getTable('recyclebin')->hasIndex('recyclebin_date')) {
+            $this->addSql('ALTER TABLE `recyclebin` DROP INDEX `recyclebin_date`');
+        }
     }
 }
