+++ bundles/CoreBundle/src/Migrations/Version20221107084519.php
@@ -29,13 +29,17 @@ final class Version20221107084519 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE object_url_slugs DROP INDEX `index`;');
-        $this->addSql('ALTER TABLE object_url_slugs DROP `index`;');
+        if($schema->hasTable('object_url_slugs') && $schema->getTable('object_url_slugs')->hasIndex('index')) {
+            $this->addSql('ALTER TABLE object_url_slugs DROP INDEX `index`;');
+            $this->addSql('ALTER TABLE object_url_slugs DROP `index`;');
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE object_url_slugs ADD `index` int(11) UNSIGNED NOT NULL DEFAULT \'0\' AFTER `fieldname`;');
-        $this->addSql('ALTER TABLE object_url_slugs ADD INDEX `index`(`index`);');
+        if($schema->hasTable('object_url_slugs') && !$schema->getTable('object_url_slugs')->hasIndex('index')) {
+            $this->addSql('ALTER TABLE object_url_slugs ADD `index` int(11) UNSIGNED NOT NULL DEFAULT \'0\' AFTER `fieldname`;');
+            $this->addSql('ALTER TABLE object_url_slugs ADD INDEX `index`(`index`);');
+        }
     }
 }
